Support for overlap, added crop markers to output PDFs
Output now defaults to a 3mm overlap on two sides to make it easier to
glue everything up afterwards. Can be specified with the -o option to
the build_map_pdfs.py tool.

Output now also includes crop markers, 5mm lines beyond the map area to
indicate corners and, when an overlap is applied, the region of that.